Update Traffic.java
Add consideration for checking the total time spent parking while calculating average speed. Not too sure if this is in the requirements, but if a vehicle is significantly lower then the average speed of their type then I should raise an alert... But if a vehicle was parked it usually means they are significantly lower then the average speed. let me know whether or not you think we need to include this