Update Traffic.java
Some error checking as well as a new inconsistency check. To add to report:
As well as checking the amount of vehicles being monitored, we also ensure that each vehicle type is consistent. I.e if a vehicle type in the statistics input is not found in the vehicle input, the system will end.